step 1: move code editor iframe entry point to its own file
